Spring data integration

+++ oauth-server/src/main/groovy/tw/edu/ncu/cc/oauth/server/helper/data/SerialSecret.groovy
@@ -2,15 +2,15 @@ package tw.edu.ncu.cc.oauth.server.helper.data;
 
 public class SerialSecret {
 
-    private long id;
+    private Integer id;
     private String secret;
 
-    public SerialSecret( long id, String secret ) {
+    public SerialSecret( Integer id, String secret ) {
         this.id = id;
         this.secret = secret;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
